Bina_yonetim/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.1

# PostgreSQL bağlantısı
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Deployment
gunicorn==22.0.0
whitenoise==6.6.0

# Extra libs (sizde var, kalsın)
pandas==2.2.2
reportlab==4.1.0
